fix(loop): :bug: Consolidate details in sidesheet to single heading
