+++ libs/loopsidesheet/src/lib/ui-sidesheet/DetailsTab.tsx
@@ -1,9 +1,13 @@
 import { Loop } from '@cc-components/loopshared';
+import { StatusCircle } from '@cc-components/shared/common';
+import { statusColorMap } from '@cc-components/shared/mapping';
 import {
   stringCell,
   StyledTabContent,
   StyledTable,
 } from '@cc-components/shared/sidesheet';
+import { formatDateString } from '@cc-components/shared/utils-dates';
+import { ContentTable } from './LoopContentTable';
 
 type DetailsTabProps = {
   loop: Loop | undefined;
@@ -33,8 +37,38 @@ export const DetailsTab = ({ loop }: DetailsTabProps): JSX.Element | null => {
             <td>Status</td>
             <td>{stringCell(loop.tagStatus)}</td>
           </tr>
+          <tr>
+            <td>MC status of content</td>
+            <td>
+              {loop.loopContentStatus ? (
+                <StatusCircle
+                  content={loop.loopContentStatus}
+                  statusColor={statusColorMap[loop.loopContentStatus]}
+                />
+              ) : (
+                '-'
+              )}
+            </td>
+          </tr>
+          <tr>
+            <td>Planned MC complete of content</td>
+            <td>
+              {loop.woPlannedCompletionDate
+                ? stringCell(formatDateString(loop.woPlannedCompletionDate))
+                : '-'}
+            </td>
+          </tr>
+          <tr>
+            <td>Actual MC complete of content</td>
+            <td>
+              {loop.woActualCompletionDate
+                ? stringCell(formatDateString(loop.woActualCompletionDate))
+                : '-'}
+            </td>
+          </tr>
         </tbody>
       </StyledTable>
+      <ContentTable loop={loop} />
     </StyledTabContent>
   );
 };
